ChangeToDownloadInPython (#3)
* add download in python

threaded download support in python. add more configs to specify download if we want to download in python and the download location and thread count

* update config_example

* add validation txt to check before deleting folder

i wanted to make sure that we check the folder contents to validate that it is a valid dramascrape downloads folder

* exit if validation text is not found

user can manually create the .txt file or change the download folder path if so

* add logger file

logger file used for validating if folder is safe to delete. it logs start time of download and finished URLs. update config_example

* update logger names

* change write to file tring